Use the glossaries package for the glossary
Latexmk with outdir/auxdir and makeglossaries do not play together nicely at all,
so currently the makefile may cause up to four calls to pdflatex.

The current setup uses makeindex for sorting, which has a bunch of drawbacks
regarding sorting of entries that contain umlauts. If this turns out to be a
problem, we will have to switch to xindy and bib2gls, but they are both not
shipped with my TeX distribution, so I will wait with that change until it
turns out to be actually necessary.