Implement new slide-in main menu
The new menu doesn't obstruct the editor window, and can safely be left
open while developing lambda terms. This allows us to simply leave the
library reference open, and allow multiple term info panels to be
expanded at the same time.
Of course, this means the slide-in menu will need to be scrollable if it
gets too large. However, this is deemed fine, as users would expect this
behavior from a slide-in menu (unlike from a pop-up)